Replace `CTimerPwmPin` with `Channel`
